+++ packages/perseus-editor/src/article-editor.tsx
@@ -4,7 +4,13 @@
  * multiple (Renderer) sections concatenated together.
  */
 
-import {components, ApiOptions, iconTrash} from "@khanacademy/perseus";
+import {
+    components,
+    ApiOptions,
+    iconTrash,
+    conversionRequired,
+    convertDeprecatedWidgets,
+} from "@khanacademy/perseus";
 import {Errors, PerseusError} from "@khanacademy/perseus-core";
 import Banner from "@khanacademy/wonder-blocks-banner";
 import * as React from "react";
@@ -20,10 +26,6 @@ import {
     iconCircleArrowUp,
     iconPlus,
 } from "./styles/icon-paths";
-import {
-    convertDeprecatedWidgets,
-    conversionRequired,
-} from "./util/deprecated-widgets/modernize-widgets-utils";
 
 import type {
     APIOptions,
@@ -240,10 +242,13 @@ export default class ArticleEditor extends React.Component<Props, State> {
                                     {...section}
                                     apiOptions={apiOptions}
                                     imageUploader={imageUploader}
-                                    onChange={_.partial(
-                                        this._handleEditorChange,
-                                        i,
-                                    )}
+                                    // TODO(LEMS-2656): remove TS suppression
+                                    onChange={
+                                        _.partial(
+                                            this._handleEditorChange,
+                                            i,
+                                        ) as any
+                                    }
                                     placeholder="Type your section text here..."
                                     ref={"editor" + i}
                                 />