+++ packages/perseus/src/__tests__/renderer.test.tsx
@@ -47,7 +47,11 @@ jest.mock("../translation-linter", () => {
 
 describe("renderer", () => {
     beforeAll(() => {
+        // TODO(LEMS-2656): remove TS suppression
+        // @ts-expect-error: InputNumberExport is not assignable to type WidgetExports
         registerWidget("numeric-input", InputNumberExport);
+        // TODO(LEMS-2656): remove TS suppression
+        // @ts-expect-error: RadioWidgetExport is not assignable to type WidgetExports
         registerWidget("radio", RadioWidgetExport);
     });