+++ packages/perseus/src/index.ts
@@ -82,6 +82,10 @@ export {isItemRenderableByVersion} from "./renderability";
 export {violatingWidgets} from "./a11y";
 export {interactiveSizes} from "./styles/constants";
 export {displaySigFigs} from "./sigfigs";
+export {
+    conversionRequired,
+    convertDeprecatedWidgets,
+} from "./util/deprecated-widgets/modernize-widgets-utils";
 
 // Context for managing i18n
 export {
@@ -101,6 +105,7 @@ export {
     containerSizeClass,
     getInteractiveBoxFromSizeClass,
 } from "./util/sizing-utils";
+export {mathOnlyParser} from "./widgets/interactive-graphs/utils";
 export {
     getAnswersFromWidgets,
     injectWidgets,