+++ packages/perseus/src/widgets/grapher/grapher.tsx
@@ -113,6 +113,8 @@ class FunctionGrapher extends React.Component<FunctionGrapherProps> {
     };
 
     change = (...args) => {
+        // TODO(LEMS-2656): remove TS suppression
+        // @ts-expect-error: Argument of type 'any[]' is not assignable to parameter of type '[newPropsOrSinglePropName: string | { [key: string]: any; }, propValue?: any, callback?: (() => unknown) | undefined]'. Target requires 1 element(s) but source may have fewer.
         return Changeable.change.apply(this, args);
     };
 
@@ -652,5 +654,7 @@ export default {
     widget: Grapher,
     transform: propTransform,
     staticTransform: staticTransform,
+    // TODO(LEMS-2656): remove TS suppression
+    // @ts-expect-error: Type 'UserInput' is not assignable to type 'PerseusGrapherUserInput'.
     scorer: scoreGrapher,
 } satisfies WidgetExports<typeof Grapher>;