+++ packages/perseus/src/widgets/interactive-graphs/mafs-state-to-interactive-graph.test.ts
@@ -271,6 +271,7 @@ describe("mafsStateToInteractiveGraph", () => {
                 [3, 4],
                 [5, 6],
             ],
+            closedPolygon: false,
         };
 
         const result: PerseusGraphType = mafsStateToInteractiveGraph(